Adds JavaDoc generation step to build
This commit adds the final functionality of the stable build, the
JavaDoc generation.  Note that it for some reason needs to be run after
the deploy step, the other way around will break the deploy step.

* We need to really recreate the softlink to the documentation git in
every build.  I'm not entirely sure why, the storage is obviously more
persistent across builds than the docker container itself should be.
The link should still be right during the next build, seeing as it is
not an absolute path name, but for some reason it doesn't work as
expected.

* The dir() encapsulation doesn't work in sh commands together with
docker containers.  There's a Jenkins bug filed on this:
  https://issues.jenkins-ci.org/browse/JENKINS-33510

* Credentials variable gets lost in the docker conversion, they are now
retrieved in the stage under withCredentials.  Turns out you can't
easily use your email as a login in bitbuckets http(s) interface, so we
have now created a new credential set in Jenkins that uses username
instead of the email.

* When it comes to paths, due to the way docker integrates with the
build, the ${PWD} variable doesn't contain the correct string until you
actually execute the command that needs the absolute path.  This has the
interesting side effect that all variables that contain a reference to
the ${PWD} variable must be unresolved (single quoted) at the time of
setting it so that it can be resolved by the shell inside the command
itself.  At the same time, a variable that contains another variable
isn't going to work in the shell as it isn't able to resolve recursively
all the variables inside other variables.  Therefore all other variables
must be resolved via double quotes at the time of including them into
other variables, so that only the ${PWD} variable remains at sh
execution time.

Example: project_path = '${PWD}/project'
	 project_lib_path = "${project_path}/build/target/libs"
In the above example, it's important that PWD isn't resolved at the time
of setting the project_path variable (under the environment tag), as jenkins isn't
executing in the temporary build space at that time.  At the same time,
it's important that project_lib_path IS resolved, so that only ${PWD}
remains as the final variable to resolve when we finally run the command
with the variable in it.

* We now want the publish step to always run, and upload SNAPSHOT
binaries to artifactory as well. Condition removed for Publish step.

Change-Id: I75d428057cb22322d2003ee7ad2d6619544fc283